change the type of the probability parameter :tada:

+++ src/main/java/com/ungs/myogenerator/commandline/CommandLineOptions.java
@@ -3,9 +3,9 @@ package com.ungs.myogenerator.commandline;
 public class CommandLineOptions {
     private final Integer size;
     private final Integer length;
-    private final Integer probability;
+    private final Float probability;
 
-    public CommandLineOptions(Integer size, Integer length, Integer probability) {
+    public CommandLineOptions(Integer size, Integer length, Float probability) {
         this.size = size;
         this.length = length;
         this.probability = probability;
